Move fairworkflows lib to core. (Quick move, not yet working)
