+++ README.md
@@ -21,3 +21,26 @@ If you just want to see what the output looks like, you can run:
 ```python generate_sample_workflow.py```
 
 This will generate a sample workflow with one step, and place the result in the folder `sample_output/`.
+
+
+
+# ```fairworkflows``` python library
+## Installation
+
+From the root of the repo (directory containing ```setup.py```), run:
+
+```
+pip install .
+```
+
+## Usage
+In your python notebook, import the library as follows:
+```
+import fairworkflows as fair
+```
+
+Try out the interactive widget:
+```
+fair.search()
+```
+
