--- core/nanopub/__init__.py
@@ -1,1 +1,0 @@
-from .nanopub import *
