--- fairworkflows/README.md
@@ -1,21 +1,0 @@
-# Installation
-
-From the ```fairworkflows``` directory containing ```setup.py```
-```
-pip install .
-```
-
-# Usage
-In your python notebook, import the library as follows:
-```
-import fairworkflows as fair
-```
-
-Try out the interactive widgets:
-```
-fair.Search()
-
-fair.Load()
-
-fair.Step()
-```
